refactor tests
All of the test files are now under `testfiles`. The input files are
in `testfiles/input` and the output files are in `testfiles/output`.
The expected and actual files are _both_ in the output directory. This
is so that they are easy to compare.

Additionally, the tests have been refactored to remove duplication.