change jdom formatter
The main change here is to change the jdom formatter so that it pretty
prints while also preserving whitespace within text elements. This
allows the whitespace to be preserved in codec codes. This replaces
the prior solution which was to add the `xml:space` attribute, because
it wasn't workable in the OTS output.

Additionally, XMLUnit no longer ignores whitespace when diffing, so a
few tweaks were necessary to get all of the tests passing again.