add Tika install
The project is reconfigured to use multiple poms. There is a parent
pom, FITS pom, and a pom for each tool that's installed. The FITS pom
is essentially the same as the old main pom. The parent pom defines
some shared values, and glues all of the child poms together. Each of
the tool poms is responsible for installing an individual tool.

This structure works well because it isolates each tools dependencies
from each other. It was proving challenging to make this work within a
single pom because Maven is not designed to isolate dependencies
within a project.

The end result here is that the Tika jars are no longer in the source
tree, and Tika can easily be upgraded simply by changing the version
number in the parent pom.